Move perfect foresight to a sub module
