+++ gEconpy/model/perfect_foresight/__init__.py
@@ -1,0 +1,11 @@
+from gEconpy.model.perfect_foresight.compile import (
+    PerfectForesightProblem,
+    compile_perfect_foresight_problem,
+)
+from gEconpy.model.perfect_foresight.solve import solve_perfect_foresight
+
+__all__ = [
+    "PerfectForesightProblem",
+    "compile_perfect_foresight_problem",
+    "solve_perfect_foresight",
+]
